Add default implementations of some AutoState methods
Most of the time, an AutoState wants to return 0 for arm motion or
intake speed, or perhaps wants to remain stationary. This was the case
for more than half of the subclasses we have, so this reduces code.

Add the @Override attribute in some places it was previously missing
